Complete implementation of parent processing.
Despite the parent nomenclature, this can be used to compare the case and
death counts of any two entities, such as comparing Arizona rates to
New York.  Usually, however, it is used to see an entity's contribution
to its parent entity, such as New York to the US, or La Paz county to
Arizona.