Do active cases right.
Just pulling them out of the file doesn't work because not all jurisdictions maintain
that field. Now computes active cases as:
cumulative confirmed cases - (cumulative deaths + confirmed cases as of 14 days ago)
The reason I chose 14 days ago is that seems to be the time it takes to be non-
contagious. For the purposes of this exercise, I deem that the case is no longer
active at that point.